feat: rename catalogues for Sentinel5P
