+++ libs/map/data-access-stac-catalog/src/lib/query-builder/fields/copernicus/copernicus.field.ts
@@ -2,7 +2,7 @@ import { TCopernicusParams, TFields } from '../../query.model';
 import { getSentinel1Fields } from './sentinel-1.field';
 import { getSentinel2Fields } from './sentinel-2.field';
 import { getSentinel3Fields } from './sentinel-3.field';
-import { getSentinel5PFields } from './sentinel-5.field';
+import { getSentinel5PFields } from './sentinel-5p.field';
 
 export const getFieldsForCopernicus = (params: TCopernicusParams): TFields => {
   switch (params.type) {
