+++ libs/map/data-access-stac-catalog/src/lib/query-builder/filter-params/copernicus/copernicus.filter-params.ts
@@ -2,7 +2,7 @@ import { TCopernicusParams, TFilterParam } from '../../query.model';
 import { createSentinel1FilterParams } from './sentinel-1/sentinel-1.filter-params';
 import { createSentinel2FilterParams } from './sentinel-2.filter-params';
 import { createSentinel3FilterParams } from './sentinel-3.filter-params';
-import { createSentinel5PFilterParams } from './sentinel-5.filter-params';
+import { createSentinel5PFilterParams } from './sentinel-5p.filter-params';
 
 export const createCopernicusParams = (params: TCopernicusParams): TFilterParam[] => {
   switch (params.type) {
